Change classes artifact jar naming
Currently, the `classes.jar` file from a source AAR is renamed to `dependencyName` and put inside the libs folder when creating the final AAR dependency. However, as it uses only the name, it may clash with another `.jar` file dependency it has and thus losing its classes.

Example: If I have a project dependency called example (`project(:example)`) and inside this project I have another sub-dependency called `example.jar`, the classes of the main project file would be lost as the sub-dependency would overwrite the jar with the same name.

This PR changes the naming of the dependency artifact to user the format `group-name-version` so we avoid naming clashes.

Co-authored-by: Igor Borges <igor@borges.me>